requirements.txt
# Django & REST
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2

# Database
psycopg2-binary>=2.9

# Utilities
python-dotenv>=1.0
django-cors-headers>=4.0
httpx>=0.24.0

# File Handling
Pillow>=10.0
PyPDF2>=3.0
PyMuPDF>=1.24

# Cloud Storage
supabase>=2.0.0

# Deployment
gunicorn>=20.1
whitenoise>=6.6
